+++ core-ui/src/components/Extensibility/useGetCRbyPath.js
@@ -1,6 +1,4 @@
 import { useMicrofrontendContext } from 'shared/contexts/MicrofrontendContext';
-import { tempDetails } from 'components/Extensibility/tempRes/tempDetails';
-import { translations } from 'components/Extensibility/tempRes/translations';
 
 export const useGetCRbyPath = () => {
   const { customResources = [], namespaceId } = useMicrofrontendContext();
@@ -18,9 +16,5 @@ export const useGetCRbyPath = () => {
     return crPath.includes(`/${urlPath}`);
   });
 
-  //resource
-  resource.details = tempDetails;
-  resource.translations = translations;
-  console.log(resource);
   return resource;
 };
