+++ core/src/luigi-config/customResources/destinationRules/form/index.js
@@ -1,0 +1,24 @@
+import { trafficPolicyBundle } from './trafficPolicy/trafficPolicyBundle';
+import { subsets } from './subsets';
+
+export const form = [
+  {
+    simple: true,
+    path: 'spec.host',
+    name: 'Host',
+  },
+  {
+    widget: 'FormGroup',
+    simple: true,
+    path: 'spec.trafficPolicy',
+    name: 'Traffic Policy',
+    children: trafficPolicyBundle,
+  },
+  subsets,
+  { path: 'spec.exportTo', name: 'Export To', widget: 'SimpleList' },
+  {
+    path: 'spec.workloadSelector.matchLabels',
+    name: 'Workload Selector / Match Labels',
+    widget: 'KeyValuePair',
+  },
+];
