+++ core/src/luigi-config/customResources/destinationRules/index.js
@@ -1,65 +1,8 @@
-import { trafficPolicyBundle } from './trafficPolicy/trafficPolicyBundle';
-import { subsets } from './subsets';
+import { form } from './form/index';
+import { details } from './details/index';
+import { translations } from './translations';
 
 export const destinationRules = {
-  details: {
-    header: [],
-    body: [
-      {
-        name: 'References',
-        widget: 'Panel',
-        children: [
-          {
-            source: 'spec.host',
-            type: 'string',
-            name: 'Host',
-          },
-          {
-            source: 'spec.exportTo',
-            widget: 'JoinedArray',
-            separator: ', ',
-          },
-        ],
-      },
-      {
-        source: 'spec.trafficPolicy',
-        name: 'spec.trafficPolicy',
-        widget: 'CodeViewer',
-        visibility: '$exists($.data)',
-      },
-      {
-        source: 'spec.subset',
-        widget: 'CodeViewer',
-        visibility: '$exists($.data)',
-      },
-      {
-        source: 'spec.workloadSelector',
-        widget: 'CodeViewer',
-        visibility: '$exists($.data)',
-      },
-    ],
-  },
-  form: [
-    {
-      simple: true,
-      path: 'spec.host',
-      name: 'Host',
-    },
-    {
-      widget: 'FormGroup',
-      simple: true,
-      path: 'spec.trafficPolicy',
-      name: 'Traffic Policy',
-      children: trafficPolicyBundle,
-    },
-    subsets,
-    { path: 'spec.exportTo', name: 'Export To', widget: 'SimpleList' },
-    {
-      path: 'spec.workloadSelector.matchLabels',
-      name: 'Workload Selector / Match Labels',
-      widget: 'KeyValuePair',
-    },
-  ],
   general: {
     resource: {
       kind: 'DestinationRule',
@@ -72,24 +15,13 @@ export const destinationRules = {
     scope: 'namespace',
     description: 'resource.description',
   },
+  details,
+  form,
   list: [
     {
       source: 'spec.host',
       name: 'Host',
     },
   ],
-  translations: {
-    en: {
-      'metadata.annotations': 'Annotations',
-      'metadata.labels': 'Labels',
-      'metadata.creationTimestamp': 'Created at',
-      'resource.description':
-        '{{[Destination Rule](https://istio.io/latest/docs/reference/config/networking/destination-rule)}} specifies rules that apply to traffic intended for a service after routing.',
-      references: 'References',
-      'spec.exportTo': 'Exported To Namespaces',
-      'spec.host': 'Host',
-      'spec.trafficPolicy': 'Traffic Policy',
-      'spec.subsets': 'Subsets',
-    },
-  },
+  translations,
 };
