+++ core/src/luigi-config/customResources/index.js
@@ -1,3 +1,3 @@
-import { destinationRules } from './destinationRules';
+import { destinationRules } from './destinationRules/index';
 
 export const busolaOwnExtConfigs = [destinationRules];
